requirements.txt
aiohttp==3.9.0
aiosignal==1.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.4
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.2.1
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.3
motor==3.3.2
multidict==6.0.4
orjson==3.9.10
packaging==23.2
pip==23.3
pluggy==1.3.0
pyaes==1.6.1
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydocstyle==6.3.0
pymongo==4.6.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.27.0
telegram-webapp-auth==1.0.1
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
yarl==1.9.3
